Functions

Index

IdealistaAPIClient.get_token

IdealistaAPIClient.process_response

IdealistaAPIClient.search

IdealistaAPIClient.search

IdealistaAPIClient.stringdict_to_nt

IdealistaAPIClient.valid_fields

IdealistaAPIClient.valid_fields

IdealistaAPIClient.get_token — Method

get_token()

Return Idealista Search API parsed response for token request

Notes

APIKEY and SECRET must be defined as environmental variables. Generated tokens get cached in tempdir, and any new calls to the get_token function before the expiration date will return the cached value

Examples

julia>get_token()
[ Info: Getting new access token
Dict{String, Any} with 5 entries:
  "access_token" => "token_value"
  "token_type"   => "bearer"
  "scope"        => "read"
  "expires_in"   => DateTime("2022-08-14T03:07:24.573")
  "jti"          => "jti_value"

IdealistaAPIClient.process_response — Method

process_response(response)

Process the response of the Idealista Search API

Goes through the dictionary of the parsed Idealista Search API response and instantiates a Response object

Examples

IdealistaAPIClient.search — Function

search(base_search::Search, property::Union{<:PropertyFields, Nothing}=nothing; token::Union{Dict{String, Any}, Nothing}=nothing)

Perform a search in the Idealista Search API

This method uses Search and property types instances as arguments

Arguments

base_search: Search object

property: Homes, Premises, Garages, Bedrooms or Offices object

Keyword Argyments

token_d: Dict containing the API response for a token request

Examples

julia>search(Search(country="es", center="40.430,-3.702", propertyType="homes", distance=15000, operation="sale"), Homes(bedrooms="1,2,3,4", swimmingPool=true))
[ Info: Getting new access token
Dict{String, Any} with 12 entries:
  "hiddenResults"      => false
  "itemsPerPage"       => 20
  "upperRangePosition" => 20
  "totalPages"         => 206
  "paginable"          => true
  "summary"            => Any["Comprar", "Viviendas", "barrio Trafalgar, Madrid", "Todos los…
  "total"              => 4118
  "lowerRangePosition" => 0
  "alertName"          => "Viviendas en barrio Trafalgar, Madrid"
  "elementList"        => Any[Dict{String, Any}("rooms"=>6, "propertyCode"=>"98324891", "num…
  "numPaginations"     => 0
  "actualPage"         => 1

IdealistaAPIClient.search — Method

search(base_search::Search, property::Union{<:PropertyFields, Nothing}=nothing; token::Union{Dict{String, Any}, Nothing}=nothing)

Perform a search in the Idealista Search API

This method uses keyword arguments corresponding to valid search fields of the Idealista Search API

Keyword Argyments

token_d: Dict containing the API response for a token request

kwargs: search fields of the Idealista Search API

Examples

julia>search(country="es", center="40.430,-3.702", propertyType="homes", distance=15000, operation="sale", bedrooms="1,2,3,4", swimmingPool=true)
[ Info: Getting new access token
Dict{String, Any} with 12 entries:
  "hiddenResults"      => false
  "itemsPerPage"       => 20
  "upperRangePosition" => 20
  "totalPages"         => 206
  "paginable"          => true
  "summary"            => Any["Comprar", "Viviendas", "barrio Trafalgar, Madrid", "Todos los…
  "total"              => 4118
  "lowerRangePosition" => 0
  "alertName"          => "Viviendas en barrio Trafalgar, Madrid"
  "elementList"        => Any[Dict{String, Any}("rooms"=>6, "propertyCode"=>"98324891", "num…
  "numPaginations"     => 0
  "actualPage"         => 1

IdealistaAPIClient.stringdict_to_nt — Method

stringdict_to_nt(::Dict{String, T})

Convert a Dict with String keys into a NamedTuple

Examples

d = Dict("a"=>2, "b"=>3)

stringdict_to_nt(d)
(a=2, b=3)

IdealistaAPIClient.valid_fields — Method

valid_fields(T::Type; indent::Int=0)

Print fieldnames of type T

Prints fieldnames of the passed type to standard output

Keyword Arguments

indent: number of tabs of indentation

Examples

julia>valid_fields(Garages)
bankOffer
automaticDoor
motorcycleParking
security

IdealistaAPIClient.valid_fields — Method

valid_fields(;indent::Int=1)

Prints fieldnames of all SearchFields subtypes

Prints to stdout names and fieldnames of subtypes of SearchFields

Keyword Arguments

indent: number of tabs of indentation

Examples

julia>valid_fields()
Search
	country
	operation
	propertyType
	center
	distance
	locationId
	maxItems
	numPage
	maxPrice
	minPrice
	sinceDate
	order
	sort
	adIds
	hasMultimedia
Bedrooms
	housemates
	smokePolicy
	petsPolicy
	gayPartners
	newGender
Garages
	bankOffer
	automaticDoor
	motorcycleParking
	security
Homes
	minSize
	maxSize
	virtualTour
	flat
	penthouse
	duplex
	studio
	chalet
	countryHouse
	bedrooms
	bathrooms
	preservation
	newDevelopment
	furnished
	bankOffer
	garage
	terrace
	exterior
	elevator
	swimmingPool
	airConditioning
	storeRoom
	clotheslineSpace
	builtinWardrobes
	subTypology
Offices
	minSize
	maxSize
	layout
	buildingType
	garage
	hotWater
	heating
	elevator
	airConditioning
	security
	exterior
	bankOffer
Premises
	minSize
	maxSize
	virtualTour
	location
	corner
	airConditioning
	smokeVentilation
	heating
	transfer
	buildingTypes
	bankOffer