IdealistaAPIClient.Bedrooms

IdealistaAPIClient.DetailedType

IdealistaAPIClient.Element

IdealistaAPIClient.Garages

IdealistaAPIClient.Homes

IdealistaAPIClient.Offices

IdealistaAPIClient.ParkingSpace

IdealistaAPIClient.Premises

IdealistaAPIClient.Response

IdealistaAPIClient.Search

Types specification

IdealistaAPIClient.Bedrooms — Type

Bedrooms <: PropertyFields

A struct that stores the bedrooms specific search fields

Each field represents a valid bedrooms search field for the Idealista Search API

Constructors

Bedrooms(housemates,
         smokePolicy,
         petsPolicy,gayPartners,
         newGender)

Bedrooms(; housemates::Union{<:AbstractString, Nothing}=nothing,
           smokePolicy::Union{<:AbstractString, Nothing}=nothing,
           petsPolicy::Union{<:AbstractString, Nothing}=nothing,
           gayPartners::Union{Bool, Nothing}=nothing,
           newGender::Union{<:AbstractString, Nothing}=nothing)

Examples

julia> Bedrooms("2,4", "disallowed", "disallowed", true, "male")
Bedrooms:
	housemates => 2,4
	smokePolicy => disallowed
	petsPolicy => disallowed
	gayPartners => true
	newGender => male

julia> Bedrooms(housemates="2,4", smokePolicy="disallowed", petsPolicy="disallowed", gayPartners=true, newGender="male")
Bedrooms:
	housemates => 2,4
	smokePolicy => disallowed
	petsPolicy => disallowed
	gayPartners => true
	newGender => male

IdealistaAPIClient.DetailedType — Type

DetailedType <: ResponseFields

A struct that stores the detailed Type of properties returned by the Idealista Search API

Constructors

DetailedType(typology, subTypology)

DetailedType(; typology::String,
               subTypology::Union{String, Nothing}=nothing)

IdealistaAPIClient.Element — Type

Element <: ResponseFields

Generic response fields for all property types

Notes

In the future there should be diferent types for each property type (i.e. Homes, Offices, Premses, Bedrooms and Garages)

IdealistaAPIClient.Garages — Type

Garages <: PropertyFields

A struct that stores the garage specific search fields

Each field represents a valid garages search field for the Idealista Search API

Constructors

Garages(bankOffer, automaticDoor, motorcycleParking, security)

Garages(; bankOffer::Union{Bool, Nothing}=nothing,
          automaticDoor::Union{Bool, Nothing}=nothing,
          motorcycleParking::Union{Bool, Nothing}=nothing,
          security::Union{Bool, Nothing}=nothing)

Examples

julia> Garages(nothing, true, true, false)
Garages:
	bankOffer => nothing
	automaticDoor => true
	motorcycleParking => true
	security => false

julia> Garages(automaticDoor=true, security=true)
Garages:
	bankOffer => nothing
	automaticDoor => true
	motorcycleParking => nothing
	security => true

IdealistaAPIClient.Homes — Type

Homes <: PropertyFields

A struct that stores the homes specific search fields

Each field represents a valid homes search field for the Idealista Search API

Constructors

Homes(minSize,
      maxSize,
      virtualTour,
      flat,
      penthouse,
      duplex,
      studio,
      chalet,
      countryHouse,
      bedrooms,
      bathrooms,
      preservation,
      newDevelopment,
      furnished,
      bankOffer,
      garage,
      terrace,
      exterior,
      elevator,
      swimmingPool,
      airConditioning,
      storeRoom,
      clotheslineSpace,
      builtinWardrobes,
      subTypology)

struct Homes(; minSize::Union{<:Number, Nothing}=nothing,
               maxSize::Union{<:Number, Nothing}=nothing,
               virtualTour::Union{Bool, Nothing}=nothing,
               flat::Union{Bool, Nothing}=nothing,
               penthouse::Union{Bool, Nothing}=nothing,
               duplex::Union{Bool, Nothing}=nothing,
               studio::Union{Bool, Nothing}=nothing,
               chalet::Union{Bool, Nothing}=nothing,
               countryHouse::Union{Bool, Nothing}=nothing,
               bedrooms::Union{<:AbstractString, Nothing}=nothing,
               bathrooms::Union{<:AbstractString, Nothing}=nothing,
               preservation::Union{<:AbstractString, Nothing}=nothing,
               newDevelopment::Union{Bool, Nothing}=nothing,
               furnished::Union{<:AbstractString, Nothing}=nothing,
               bankOffer::Union{Bool, Nothing}=nothing,
               garage::Union{Bool, Nothing}=nothing,
               terrace::Union{Bool, Nothing}=nothing,
               exterior::Union{Bool, Nothing}=nothing,
               elevator::Union{Bool, Nothing}=nothing,
               swimmingPool::Union{Bool, Nothing}=nothing,
               airConditioning::Union{Bool, Nothing}=nothing,
               storeRoom::Union{Bool, Nothing}=nothing,
               clotheslineSpace::Union{Bool, Nothing}=nothing,
               builtinWardrobes::Union{Bool, Nothing}=nothing,
               subTypology::Union{<:AbstractString, Nothing}=nothing)

Examples

julia> Homes(65, 130, false, true, true, true, true, true, true, "1,2,3,4", "1,2", "good", true, "furnished", false, true, true, false, false, false, false, false, false, true, "independantHouse")
[ Info: bankOffer only applies in Spain
Homes:
	minSize => 65
	maxSize => 130
	virtualTour => false
	flat => true
	penthouse => true
	duplex => true
	studio => true
	chalet => true
	countryHouse => true
	bedrooms => 1,2,3,4
	bathrooms => 1,2
	preservation => good
	newDevelopment => true
	furnished => furnished
	bankOffer => false
	garage => true
	terrace => true
	exterior => false
	elevator => false
	swimmingPool => false
	airConditioning => false
	storeRoom => false
	clotheslineSpace => false
	builtinWardrobes => true
	subTypology => independantHouse

julia> Homes(minSize=70, swimmingPool=false, preservation="renew", subTypology="terracedHouse")
Homes:
	minSize => 70
	maxSize => nothing
	virtualTour => nothing
	flat => nothing
	penthouse => nothing
	duplex => nothing
	studio => nothing
	chalet => nothing
	countryHouse => nothing
	bedrooms => nothing
	bathrooms => nothing
	preservation => renew
	newDevelopment => nothing
	furnished => nothing
	bankOffer => nothing
	garage => nothing
	terrace => nothing
	exterior => nothing
	elevator => nothing
	swimmingPool => false
	airConditioning => nothing
	storeRoom => nothing
	clotheslineSpace => nothing
	builtinWardrobes => nothing
	subTypology => terracedHouse

IdealistaAPIClient.Offices — Type

Offices <: PropertyFields

A struct that stores the office specific search fields

Each field represents a valid offices search field for the Idealista Search API

Constructors

Offices(minSize,
        maxSize,
        layout,
        buildingType,
        garage,
        hotWater,
        heating,
        elevator,
        airConditioning,
        security,
        exterior,
        bankOffer)

Offices(; minSize::Union{<:Number, Nothing}=nothing,
          maxSize::Union{<:Number, Nothing}=nothing,
          layout::Union{<:AbstractString, Nothing}=nothing,
          buildingType::Union{<:AbstractString, Nothing}=nothing,
          garage::Union{Bool, Nothing}=nothing,
          hotWater::Union{Bool, Nothing}=nothing,
          heating::Union{Bool, Nothing}=nothing,
          elevator::Union{Bool, Nothing}=nothing,
          airConditioning::Union{Bool, Nothing}=nothing,
          security::Union{Bool, Nothing}=nothing,
          exterior::Union{Bool, Nothing}=nothing,
          bankOffer::Union{Bool, Nothing}=nothing)

Examples

julia> Offices(100, 400, "withWalls", "exclusive", false, true, true, true, true, false, false, false)
[ Info: bankOffer only applies in Spain
Offices:
	minSize => 100
	maxSize => 400
	layout => withWalls
	buildingType => exclusive
	garage => false
	hotWater => true
	heating => true
	elevator => true
	airConditioning => true
	security => false
	exterior => false
	bankOffer => false

julia> Offices(minSize=100, maxSize=400, layout="withWalls", buildingType="exclusive")
Offices:
	minSize => 100
	maxSize => 400
	layout => withWalls
	buildingType => exclusive
	garage => nothing
	hotWater => nothing
	heating => nothing
	elevator => nothing
	airConditioning => nothing
	security => nothing
	exterior => nothing
	bankOffer => nothing

IdealistaAPIClient.ParkingSpace — Type

ParkingSpace <: ResponseFields

A struct that stores the details of parking spaces returned by the Idealista Search API

Constructors

ParkingSpace(hasParkingSpace, isParkingSpaceIncludedInPrice, parkingSpacePrice)

ParkingSpace(; hasParkingSpace::Bool,
               isParkingSpaceIncludedInPrice::Union{Bool, Nothing}=nothing,
               parkingSpacePrice::Union{Number, Nothing}=nothing)

IdealistaAPIClient.Premises — Type

Premises <: PropertyFields

A struct that stores the premises specific search fields

Each field represents a valid premises search field for the Idealista Search API

Constructors

Premises(minSize,
         maxSize,
         virtualTour,
         location,
         corner,
         airConditioning,
         smokeVentilation,
         heating,
         transfer,
         buildingTypes,
         bankOffer)

Premises(; minSize::Union{<:Number, Nothing}=nothing,
           maxSize::Union{<:Number, Nothing}=nothing,
           virtualTour::Union{Bool, Nothing}=nothing,
           location::Union{<:AbstractString, Nothing}=nothing,
           corner::Union{Bool, Nothing}=nothing,
           airConditioning::Union{Bool, Nothing}=nothing,
           smokeVentilation::Union{Bool, Nothing}=nothing,
           heating::Union{Bool, Nothing}=nothing,
           transfer::Union{Bool, Nothing}=nothing,
           buildingTypes::Union{<:AbstractString, Nothing}=nothing,
           bankOffer::Union{Bool, Nothing}=nothing)

Examples

julia> Premises(70, 120, false, "shoppingcenter", false, true, false, true, false, "industrialBuilding", false)
[ Info: bankOffer only applies in Spain
Premises:
	minSize => 70
	maxSize => 120
	virtualTour => false
	location => shoppingcenter
	corner => false
	airConditioning => true
	smokeVentilation => false
	heating => true
	transfer => false
	buildingTypes => industrialBuilding
	bankOffer => false

julia> Premises(minSize=70, maxSize=120, transfer=false, location="shoppingcenter")
Premises:
	minSize => 70
	maxSize => 120
	virtualTour => nothing
	location => shoppingcenter
	corner => nothing
	airConditioning => nothing
	smokeVentilation => nothing
	heating => nothing
	transfer => false
	buildingTypes => nothing
	bankOffer => nothing

IdealistaAPIClient.Response — Type

Response <: ResponseFields

A struct that stores the Idealista Search API response fields

Constructors

Response(actualPage,
         itemsPerPage,
         lowerRangePosition,
         upperRangePosition,
         paginable,
         numPaginations,
         summary,
         total,
         totalPages,
         elementList,
         alertName,
         hiddenResults)

Response(; actualPage::Int64,
           itemsPerPage::Int64,
           lowerRangePosition::Int64,
           upperRangePosition::Int64,
           paginable::Bool,
           numPaginations::Int64,
           summary::Vector{String},
           total::Int64,
           totalPages::Int64,
           elementList::Vector{Element},
           alertName::String,
           hiddenResults::Bool)

IdealistaAPIClient.Search — Type

Search <: SearchFields

A struct that stores the basic (non-property-type specific) search fields of the Idealista Search API

Each field represents a valid base search field for the Idealista Search API. The country, operation, propertyType and center + distance or center + locationId are mandatory

Constructors

Search(country,
       operation,
       propertyType,
       center,
       distance,
       locationId,
       maxItems,
       numPage,
       maxPrice,
       minPrice,
       sinceDate,
       orderg,
       sort,
       adIds,
       hasMultimedia)

Search(; country::String,
         operation::String,
         propertyType::String,
         center::String,
         distance::Union{<:Int, Nothing}=nothing,
         locationId::Union{<:AbstractString, Nothing}=nothing,
         maxItems::Union{<:Int, Nothing}=nothing,
         numPage::Union{<:Int, Nothing}=nothing,
         maxPrice::Union{<:Number, Nothing}=nothing,
         minPrice::Union{<:Number, Nothing}=nothing,
         sinceDate::Union{<:AbstractString, Nothing}=nothing,
         order::Union{<:AbstractString, Nothing}=nothing,
         sort::Union{<:AbstractString, Nothing}=nothing,
         adIds::Union{<:Int, Nothing}=nothing,
         hasMultimedia::Union{Bool, Nothing}=nothing)

Examples

julia> Search("es", "sale", "homes", "42.0,-3.7", 15000, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing)
Search:
	country => es
	operation => sale
	propertyType => homes
	center => 42.0,-3.7
	distance => 15000
	locationId => nothing
	maxItems => nothing
	numPage => nothing
	maxPrice => nothing
	minPrice => nothing
	sinceDate => nothing
	order => nothing
	sort => nothing
	adIds => nothing
	hasMultimedia => nothing

julia> Search(country="it", operation="sale", propertyType="homes", center="42.0,-3.7", distance=15000)
Search:
	country => it
	operation => sale
	propertyType => homes
	center => 42.0,-3.7
	distance => 15000
	locationId => nothing
	maxItems => nothing
	numPage => nothing
	maxPrice => nothing
	minPrice => nothing
	sinceDate => nothing
	order => nothing
	sort => nothing
	adIds => nothing
	hasMultimedia => nothing